backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
numpy==1.26.4
scipy==1.13.1
soundfile==0.12.1
faiss-cpu==1.8.0.post1
pypdf==5.1.0
python-docx==1.1.2
python-pptx==0.6.23
langchain-text-splitters==0.3.2
openai-whisper==20240930
httpx==0.27.2
